requirements.txt
# Core web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.5.0

# Google AI and Cloud services (using individual packages instead of google-adk)
google-generativeai>=0.3.2
google-cloud-aiplatform[agent-engines]>=1.112.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Data processing and analysis
pandas>=2.1.4
numpy>=1.24.4
yfinance>=0.2.28

# HTTP requests and API calls
requests>=2.32.4
httpx>=0.28.1

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Database (optional - SQLite for local development)
sqlalchemy>=2.0.23
alembic>=1.13.1

# Additional utilities
python-dateutil>=2.9.0.post0
pytz>=2023.3
typing-extensions>=4.11.0

# Email processing and newsletter parsing
google-api-python-client>=2.157.0
beautifulsoup4>=4.12.2
html2text>=2020.1.16
email-validator>=2.1.0

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
